pyproject.toml
[tool.poetry]
name = "osparc-control"
version = "0.0.1"
description = "Osparc Control"
authors = ["Andrei Neagu <neagu@itis.swiss>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ITISFoundation/osparc-control"
repository = "https://github.com/ITISFoundation/osparc-control"
documentation = "https://itisfoundation.github.io/osparc-control/"
classifiers = [
    "Development Status :: 3 - Alpha",
]

[tool.poetry.urls]
Changelog = "https://github.com/ITISFoundation/osparc-control/releases"

[tool.poetry.dependencies]
python = "^3.6.2"
click = "^8.0.1"
pyzmq = ">=14.0.0"
u-msgpack-python = ">=2.7.1"
pydantic = ">=1.8.2"
tenacity = ">=8.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = {extras = ["toml"], version = "^6.1"}
safety = "^1.10.3"
mypy = "^0.910"
typeguard = "^2.13.2"
xdoctest = {extras = ["colors"], version = "^0.15.10"}
pre-commit = "^2.15.0"
flake8 = "^4.0.1"
black = ">=21.10b0"
flake8-bandit = "^2.1.2"
flake8-bugbear = "^21.9.2"
flake8-docstrings = "^1.6.0"
flake8-rst-docstrings = "^0.2.3"
pep8-naming = "^0.12.1"
darglint = "^1.8.1"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.0.1"
Pygments = "^2.10.0"
pyupgrade = "^2.29.1"
furo = ">=2021.11.12"
rope = "^0.23.0"


[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.run]
branch = true
source = ["osparc_control", "tests"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover"
]
show_missing = true
fail_under = 99

[tool.mypy]
strict = true
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
